Moved functions to utils.go
